geocoding_service/requirements.txt
aiohttp==3.6.2
requests
psycopg2-binary==2.8.4
pytest
